lexer_lex

+++ inc/tools.h
@@ -6,7 +6,8 @@
 // STRING FUNCTIONS
 bool		minishell_strchr(const char *s, int c);
 uint32_t	minishell_strlen(const char *s);
-uint64_t	minishell_strlcpy(char *dst, const char *src, size_t dstsize);
+uint64_t	minishell_strlcpy(char *dst, const char *src,
+				uint64_t dstsize);
 
 // MEMORY UTILS
 void		*minishell_memset(void *b, int c, size_t len);
